slackex v0.0.1 Slackex.Channels

Get info on your team’s Slack channels, create or archive channels, invite users, set the topic and purpose, and mark a channel as read.

Functions

archive(channel)

This method archives a channel.

create(name)

This method is used to create a channel.

history(channel, options \\ %{})

This method returns a portion of messages/events from the specified channel. To read the entire history for a channel, call the method with no latest or oldest arguments, and then continue paging using the instructions below.

info(channel)

invite(channel, user)

This method is used to invite a user to a channel. The calling user must be a member of the channel.

join(name)

This method is used to join a channel. If the channel does not exist, it is created.

kick(channel, user)

This method allows a user to remove another member from a team channel.

leave(channel)

This method is used to leave a channel.

list()

This method returns a list of all channels in the team. This includes channels the caller is in, channels they are not currently in, and archived channels but does not include private channels. The number of (non-deactivated) members in each channel is also returned.

To retrieve a list of private channels, use groups.list

mark(channel, timestamp)

This method moves the read cursor in a channel.

rename(channel, new_name)

This method renames a team channel.

The only people who can rename a channel are team admins, or the person that originally created the channel. Others will receive a “not_authorized” error.

set_purpose(channel, purpose)

This method is used to change the purpose of a channel. The calling user must be a member of the channel.

set_topic(channel, topic)

This method is used to change the topic of a channel. The calling user must be a member of the channel.

unarchive(channel)

This method unarchives a channel. The calling user is added to the channel.