slackex v0.0.1 Slackex.DoNotDisturb

Adjust and view Do Not Disturb settings for team members

Functions

end_dnd()

Ends the user’s currently scheduled Do Not Disturb session immediately.

end_snooze()

Ends the current user’s snooze mode immediately.

info(options \\ %{})

Provides information about a user’s current Do Not Disturb settings.

set_snooze(num_minutes)

Adjusts the snooze duration for a user’s Do Not Disturb settings. If a snooze session is not already active for the user, invoking this method will begin one for the specified duration.

team_info(options \\ %{})

Provides information about the current Do Not Disturb settings for users of a Slack team.