slackex v0.0.1 Slackex.Files

Get info on files uploaded to Slack, upload new files to Slack.

Functions

delete(file)

This method deletes a file from your team.

info(file, options \\ %{})

This method returns information about a file in your team.

list(options \\ %{})

This method returns a list of files within the team. It can be filtered and sliced in various ways.

upload(file, options \\ %{})

This method allows you to create or upload an existing file.