slackex v0.0.1 Slackex.Files.Comments

Handles comments on files

Functions

add(file, comment)

Add a comment to an existing file.

delete(file, id)

Delete an existing comment on a file. Only the original author of the comment or a team administrator may delete a file comment.

edit(file, id, comment)

Edit an existing comment on a file. Only the user who created a comment may make edits. Teams may configure a limited time window during which file comment edits are allowed.