slackex v0.0.1 Slackex.Groups

Get info on your team’s private channels.

Functions

archive(channel)

This method archives a private channel.

close(channel)

This method closes a private channel.

create(name)

This method creates a private channel.

create_child(channel)

This method takes an existing private channel and performs the following steps:

Renames the existing private channel (from “example” to “example-archived”).

Archives the existing private channel.

Creates a new private channel with the name of the existing private channel.

Adds all members of the existing private channel to the new private channel.

This is useful when inviting a new member to an existing private channel while hiding all previous chat history from them. In this scenario you can call groups.createChild followed by groups.invite.

The new private channel will have a special parent_group property pointing to the original archived private channel. This will only be returned for members of both private channels, so will not be visible to any newly invited members.

history(channel, options \\ %{})

This method returns a portion of messages/events from the specified private channel. To read the entire history for a private channel, call the method with no latest or oldest arguments, and then continue paging using the instructions below.

info(channel)

This method returns information about a private channel.

invite(channel, user)

This method is used to invite a user to a private channel. The calling user must be a member of the private channel.

To invite a new member to a private channel without giving them access to the archives of the private channel, call the groups.createChild method before inviting.

kick(channel, user)

This method allows a user to remove another member from a private channel.

leave(channel)

This method is used to leave a private channel.

list(options \\ %{})

This method returns a list of private channels in the team that the caller is in and archived groups that the caller was in. The list of (non-deactivated) members in each private channel is also returned.

mark(channel, timestamp)

This method moves the read cursor in a private channel.

open(channel)

This method opens a private channel.

rename(channel, name)

This method renames a private channel.

set_purpose(channel, purpose)

This method is used to change the purpose of a private channel. The calling user must be a member of the private channel.

set_topic(channel, topic)

This method is used to change the topic of a private channel. The calling user must be a member of the private channel.

unarchive(channel)

This method unarchives a private channel.