slackex v0.0.1 Slackex.IM

Get info on your direct messages.

Functions

close(channel)

This method closes a direct message channel.

history(channel, options \\ %{})

This method returns a portion of messages/events from the specified direct message channel. To read the entire history for a direct message channel, call the method with no latest or oldest arguments, and then continue paging using the instructions below.

list()

This method returns a list of all im channels that the user has.

mark(channel, timestamp)

This method moves the read cursor in a direct message channel.

open(user)

This method opens a direct message channel with another member of your Slack team.