slackex v0.0.1 Slackex.Team

Get info on your team

Functions

access_logs(options \\ %{})

This method is used to get the access logs for users on a team.

info()

This method provides information about your team.

integration_logs(options \\ %{})

This method lists the integration activity logs for a team, including when integrations are added, modified and removed. This method can only be called by Admins.