slackex v0.0.1 Slackex.UserGroups

Get info on your team’s user groups.

Functions

create(name, options)

This method is used to create a user group.

disable(usergroup, options \\ %{})

This method disables an existing user group.

enable(usergroup, options \\ %{})

This method enables a user group which was previously disabled.

list(options \\ %{})

This method returns a list of all user groups in the team. This can optionally include disabled user groups.

update(usergroup, options \\ %{})

This method updates the properties of an existing user group.