slackex v0.0.1 Slackex.Users

Get info on members of your Slack team.

Functions

get_presence(user)

This method lets you find out information about a user’s presence. Consult the presence documentation for more details.

info(user)

This method returns information about a team member.

list(options \\ %{})

This method returns a list of all users in the team. This includes deleted/deactivated users.

set_active()

This method lets the slack messaging server know that the authenticated user is currently active. Consult the presence documentation for more details.

set_presence(presence)

This method lets you set the calling user’s manual presence. Consult the presence documentation for more details.